% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{app_sys}
\alias{app_sys}
\title{Function for Accesses Files in the Current Package}
\usage{
app_sys(...)
}
\arguments{
\item{...}{Character vectors, specifying subdirectory and file(s) within the package.}
}
\description{
Function for Accesses Files in the Current Package
}
