% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{choices}
\alias{choices}
\title{Choice List}
\usage{
choices()
}
\value{
A \code{ElectProtect_choices} object.
}
\description{
An S3 vector class (subclass of list) containing the inputs,
i.e., column names, for the dashboard input sidebars. The column names
are grouped together.
}
