% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Crosstab}
\usage{
crosstab(var_ind, var_dep, reverse, weight = "weight", db)
}
\arguments{
\item{var_ind}{A length-one character vector for the independent variable.}

\item{var_dep}{A length-one character vector for the dependent variable.}

\item{reverse}{A length-one character vector indicating whether to reverse the order of the variables.}

\item{weight}{A length-one character vector for the weight variable.}

\item{db}{A sqlite (\code{S4}) database connection object.}
}
\value{
A \code{datatables} (subclass of \code{htmlwidget}) object.
}
\description{
Function for generating cross-tabulations.
}
