% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-table-functions.R
\name{design_effect}
\alias{design_effect}
\title{Design effect}
\usage{
design_effect(weight)
}
\arguments{
\item{weight}{A double vector of weights.}
}
\value{
A length-one double vector.
}
\description{
Formula taken from \href{https://www.pewresearch.org/internet/2010/04/27/methodology-85/}{link}.
}
\keyword{internal}
