% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Add directory of static resources to Shiny's web server}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{A character vector giving the library directory where the package defining the namespace was found.}

\item{pkgname}{A character svector giving the name of the package.}
}
\description{
Add directory of static resources to Shiny's web server
}
