% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectProtect-package.R
\docType{package}
\name{ElectProtect-package}
\alias{ElectProtect}
\alias{ElectProtect-package}
\title{ElectProtect: A survey dashboard on fair election}
\description{
A dashboard functioning as a user interface for analyzing survey results.
}
\author{
\strong{Maintainer}: Yang Wu \email{yangwu2020@gmail.com} (\href{https://orcid.org/0000-0001-9847-0112}{ORCID})

}
\keyword{internal}
