% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret_xtab}
\alias{interpret_xtab}
\title{Interpret (Cross-tabulation)}
\usage{
interpret_xtab(var_ind, var_dep, reverse)
}
\arguments{
\item{var_ind}{A length-one character vector for the independent variable.}

\item{var_dep}{A length-one character vector for the dependent variable.}

\item{reverse}{A length-one character vector indicating whether to reverse the order of the variables.}
}
\value{
A \code{shiny.tag.list} object.
}
\description{
Function for generating template interpretations for the cross-tabulations.
}
