% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-table-functions.R
\name{moe_design_effect}
\alias{moe_design_effect}
\title{Design effect with margin of error}
\usage{
moe_design_effect(percent, deff, n, zscore = 1.96)
}
\arguments{
\item{percent}{A double vector of percentages.}

\item{deff}{A length-one double vector.}

\item{n}{A double vector of unweighted sample sizes.}

\item{zscore}{A length-one double vector of zscore.}
}
\value{
A length-one double vector.
}
\description{
Formula taken from \href{https://www.pewresearch.org/internet/2010/04/27/methodology-85/}{link}.
}
\keyword{internal}
