% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-table-functions.R
\name{multiselect_helper}
\alias{multiselect_helper}
\title{Helper for handling multiple selection questions}
\usage{
multiselect_helper(df, var, weight, group, cols)
}
\arguments{
\item{df}{A data frame.}

\item{var}{A length-one character vector for removing prefixes.}

\item{weight}{A length-one character vector for the weight variable.}

\item{group}{A group variable for determining which code path to execute.}

\item{cols}{A character vector of multiple-selection column names.}
}
\value{
A double vector.
}
\description{
Helper for handling multiple selection questions
}
\keyword{internal}
