% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_topline}
\alias{plot_topline}
\title{Plot (Topline)}
\usage{
plot_topline(data, var)
}
\arguments{
\item{data}{A \code{datatable} object.}

\item{var}{A length-one character vector for the topline variable.}
}
\value{
A \code{plotly} (subclass of \code{htmlwidget}) object.
}
\description{
Bar chart for toplines.
}
