% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_xtab}
\alias{plot_xtab}
\title{Plot (Crosstab)}
\usage{
plot_xtab(data)
}
\arguments{
\item{data}{A \code{datatable} object.}
}
\value{
A \code{plotly} (subclass of \code{htmlwidget}) object.
}
\description{
Grouped bar chart function for cross-tabulations.
}
