% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_multi_select}
\alias{rank_multi_select}
\title{Function for computing ranked DataFrame}
\usage{
rank_multi_select(df, cols, model_col, multiselect_cat)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{A character vector of multiple selection column names to include in the ranking.}

\item{model_col}{A length-one character vector for the modeled column.}

\item{multiselect_cat}{A length-one character vector for the "selected" option for each multiple-selection column.}
}
\value{
A data frame.
}
\description{
Function to rank multiple-selection columns after grouping by a modeled column.
}
\keyword{internal}
