% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dbConnect}
\alias{dbDisconnect}
\alias{SQLite}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DBI}{\code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbDisconnect}}}

  \item{RSQLite}{\code{\link[RSQLite]{SQLite}}}
}}

