% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Run application}
\usage{
runApp(choice_list, question_list, db)
}
\arguments{
\item{choice_list}{A \code{ElectProtect_choices} object instantiated with \code{ElectProtect::choices}.}

\item{question_list}{A \code{ElectProtect_questions} object instantiated with \code{ElectProtect::questions}.}

\item{db}{A sqlite (\code{S4}) database connection object.}
}
\description{
Function to run the shiny application.
}
