% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_format}
\alias{str_format}
\title{Remove Underscore and Convert to Title Case}
\usage{
str_format(str)
}
\arguments{
\item{str}{A length-one character vector to be formatted.}
}
\value{
A length-one character vector.
}
\description{
Function for formatting column names for display.
}
