% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline.R
\name{topline}
\alias{topline}
\title{Topline}
\usage{
topline(var, weight = "weight", db)
}
\arguments{
\item{var}{A length-one character vector for the topline variable.}

\item{weight}{A length-one character vector for the weight variable.}

\item{db}{A sqlite (\code{S4}) database connection object.}
}
\value{
A \code{datatables} (subclass of \code{htmlwidget}) object.
}
\description{
Function for generating toplines.
}
