% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-table-functions.R
\name{topline_internal}
\alias{topline_internal}
\title{Topline}
\usage{
topline_internal(df, variable, weight)
}
\arguments{
\item{df}{A data frame.}

\item{variable}{A length-one character vector for the topline variable.}

\item{weight}{A length-one character vector for the weight variable.}
}
\value{
A \code{data.table} object.
}
\description{
Internal topline function for non-multiple-selection questions.
}
\keyword{internal}
