% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-table-functions.R
\name{topline_multiselect_internal}
\alias{topline_multiselect_internal}
\title{Topline (multiple-selection)}
\usage{
topline_multiselect_internal(df, var, weight)
}
\arguments{
\item{df}{A data frame.}

\item{var}{A length-one character vector for removing prefixes.}

\item{weight}{A length-one character vector for the weight variable.}
}
\value{
A \code{data.table} object.
}
\description{
Internal topline function for multiple-selection questions.
}
\keyword{internal}
